A lot of progress very happy

+++ ft_strncmp.c
@@ -6,15 +6,15 @@
 /*   By: otanovic <otanovic@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/21 16:57:11 by otanovic          #+#    #+#             */
-/*   Updated: 2024/11/22 14:04:12 by otanovic         ###   ########.fr       */
+/*   Updated: 2024/11/24 14:08:43 by otanovic         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include	<stddef.h>
+#include <stddef.h>
 
 int	strncmp(const char *s1, const char *s2, size_t n)
 {
-	int i;
+	size_t	i;
 
 	i = 0;
 	while (i < n && s1[i] != '\0' && s2[i] != '\0')
