+++ memmove.c
@@ -6,7 +6,7 @@
 /*   By: otanovic <otanovic@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/20 17:49:36 by otanovic          #+#    #+#             */
-/*   Updated: 2024/11/23 16:23:01 by otanovic         ###   ########.fr       */
+/*   Updated: 2024/11/24 15:34:30 by otanovic         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ void	*ft_memmove(void *dest, const void *src, size_t num)
 {
 	unsigned char	*d;
 	unsigned char	*s;
-	int				i;
+	size_t			i;
 
 	if (dest == src || num == 0)
 	{
